requirements.txt
pygame-ce>=2.5.0
PyOpenGL>=3.1.0
PyOpenGL-accelerate>=3.1.0
gphoto2>=2.5.0
opencv-python>=4.10.0
numpy>=2.1.3
Pillow>=11.0.0
